linux/ena: Initialize RSS components during queue count change flow
ethtool commands are available even if the interface is in down state.
The command to set the channels (`ethtool -L`) destroys the RSS
components by calling `ena_update_queue_count` -> `ena_com_rss_destroy`.
If the interface is not in UP state, then there's no re-initialization
of the RSS components until the interface is set to UP.
This leads to a scenario where RSS components are freed, but commands
such as `ethtool -x` that access RSS components can still be invoked.
Such a sequence causes a kernel panic on a NULL pointer dereference.

Sequence:
1. sudo ip link set dev <name> down
2. sudo ethtool -L <name> combined <channels>
3. sudo ethtool -x <name>

This patch adds a configuration of the RSS components after they're
being destroyed in `ena_update_queue_count` in case the interface was
not in UP state. In case the interface was in UP state during the
queue count update, then the configuration will take place through
`ena_open` sequence.

Note:
In case the interface is in DOWN state during the queue count update
routine and then set to UP state later on, `ena_rss_configure` will
be called twice. The impact would be three admin commands which set
the indirection table, hash function and hash control. Though the
commands are redundant, it avoids impacting other flows that are
calling `ena_rss_configure` such as the reset flow (through `ena_up`).

Fixes: d7c11ce96a9a ("linux/ena update ENA linux driver to version 2.2.0")
Signed-off-by: David Arinzon <darinzon@amazon.com>